pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "markov-bot"
version = "0.1.0"
description = ""
authors = [ "Ari Taylor <ari@aricodes.net>",]

[tool.black]
line-length = 95
target-version = [ "py39",]

[tool.poetry.dependencies]
python = "^3.9"
peewee = "^3.14.4"
twitch-python = "^0.0.19"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"

[tool.poetry.scripts]
mb = 'markov_bot.cli:main'
